Homer: update our base config file to include 'selectors' for nokia
As part of the recent update to Homer it will look for this data
structure 'selectors' in the YAML config to tell it what devices
from Netbox should use non-default config rendor or transport.

For now this is as simple as we use Python to generate and JSON-RPC
to push the config for any Nokia devices.

Bug: T402511
Change-Id: I0af5797f6330e3f4750c5d723911db0306decf97