sshd: use the default KexAlgorithms algorithms
Starting with ssh 10.1 a warning is thrown when connecting if the ssh
server does not support a "post-quantum" algorithm,
https://www.openssh.com/pq.html. It would be nice to avoid this warning
as people upgrade to 10.1. One of the algorithms,
sntrup761x25519-sha512, is supported in Bookworm's version
1:9.2p1-2+deb12u7 and is part of the default list.

We planned to revert to the default in Trixie, but it seems worthwhile
to do so now in Bookworm to avoid the warning, unless the concerns in
commit f73786e3afb199f0fcc7aa00544b7c44beda91c7, which I believe
originated our current list, are still considered best practice.

Hosts: O:bastionhost
Change-Id: I41078d97581ad08a3d130e1cb6fe44d5ef05b107