hcaptcha_proxy: Select the custom nginx provider instead of extras
Starting with the version of nginx in Bookworm, nginx has switched away
from the previous (and somewhat random) selection of modules built
into nginx-light, nginx-full and nginx-extras.

Instead there's now a single nginx binary package and the specific
additional modules needed can be installed as libnginx-mod-foo.

Right now hcaptcha install nginx-extras, which in addition to
the necessaty Lua module also installs a lot of unnecessary one:
- libnginx-mod-http-geoip
- libnginx-mod-http-image-filter
- libnginx-mod-http-perl
- libnginx-mod-http-xslt-filter
- libnginx-mod-mail
- libnginx-mod-stream
- libnginx-mod-stream-geoip
- libnginx-mod-http-auth-pam
-.libnginx-mod-http-cache-purge
- libnginx-mod-http-dav-ext
- libnginx-mod-http-echo
- libnginx-mod-http-fancyindex
- libnginx-mod-http-geoip2
- libnginx-mod-http-headers-more-filter
- libnginx-mod-http-subs-filter
- libnginx-mod-http-uploadprogress
- libnginx-mod-http-upstream-fair
- libnginx-mod-nchan
- libnginx-mod-stream-geoip2

Instead set the nginx variant to 'custom', so that only libnginx-mod-lua
is pulled in (the other ones can be uninstalled after Puppet as run).

Bug: T405631
Change-Id: I15eed4482a077d3785c04da2923f7b264c52cc4f