profile::amd_gpu: use a system user for the GPU node labeller
The system user uid/gid is reserved in puppet, meanwhile in
the debian package we'll make sure that the user gets added
to render/video.

The idea is to create a user that can only read from /dev the GPU's
metadata and report it to the k8s api.

Related patch:
https://gerrit.wikimedia.org/r/c/operations/debs/amd-k8s-device-plugin/+/1194942

Bug: T373806
Change-Id: I814fefcb72fc58c52d6e47afd85462db8bf07a37