Update the permissions for the dse-k8s-csi user in codfw
The permissions for this user match those in eqiad, but we had decided
in https://wikimedia.slack.com/archives/C055QGPTC69/p1758112587918419 to
use `dse-k8s-csi-rbd-ssd` as the name for the pool to be used by the RBD
csi polugin in codfw.

The corresponding pool in eqiad was named `dse-k8s-csi-ssd` before we
had decided to enable support for cephfs and we don't want to rename it,
for now.

This commit updates the caps for the dse-k8s-csi user in codfw to allow
it to access the correct pool.

Hosts: cephosd2001.codfw.wmnet
Bug: T404576
Change-Id: I37ea0150a3ec25f48ca6804569f7764c82648120