Update the definition of @dse_kubepods_networks
We have recently deployed a new dse-k8s-codfw cluster but while doing
so, it seems that the network data for the kubepods and kubesvc ranges
were inadvertently put within the data structure for eqiad.

This commit puts the network definitions in the correct place and
updated the network::constants data to include the slice from codfw.

Hosts: cephosd2001.codfw.wmnet, dse-k8s-worker2001.codfw.wmnet
Bug: T404576
Change-Id: I45bbcf1588b873270f86f06689c7eccc920f06c8