analytics::launcher: Fix the ensure parameter on the drop_event timer
We have created a new an-launcher host and we wish to start migrating
the workload acrosss from an-launcher1002 to an-launcher1003.

However, we have observed that the 'drop_event` service is attempting to
run on both of the hosts. THis is caused by a missing 'ensure' parameter
on the systemd timer.

Hosts: an-launcher1002.eqiad.wmnet,an-launcher1003.eqiad.wmnet
Bug: T402943
Change-Id: I42af4a0e16af06c63ba7a2c6580ea91a358e294f