Enable canary events on an-launcher1003
This is the first production workload on an-launcher1003.

It will be a duplicate of the job running on an-launcher1002, because an
excess of canary events is not a problem.

Subsequent refinery job migrations will switch which host runs the job,
but this one can run on both.

Hosts: an-launcher1003.eqiad.wmnet
Bug: T402943
Change-Id: I4f87284ab31955f6a455c71668385651bdac3802