Pin the version of opensearch wherever it is installed
We currently only have two versions of opensearch in use, but we are
preparing to have a wider choice of versions available due to a
planned mirroring of the upstream repositories.

As part of this preparation, we wish to ensure that our opensearch
packages will not be updated to new versions automatically, or as the
result of server reimages etc.

This commit changes the handling of the version parameter in the
opensearch manifests so that it takes a SemVer type and passes this
through to the package resource.

Hosts: P:opensearch::server
Bug: T407199
Change-Id: I49636c4cd7ffd32afda6565d58f168e0a8a1d596