Pin the version of opensearch-dashboards wherever it is used
We are currently preparing to mirror the upstream opensearch package
repositories. This means that we will automatically have all of the
versions of the opensearch packages available for install, including
opensearch-dashboards.

For this reason, we wish to pin the version of opensearch-dashboards
that gets installed. We use a SemVer type for the $version parameter and
pass this through to the package resource.

Hosts: P:opensearch::dashboards
Bug: T407199
Change-Id: Idfd5dc7c97a61352678fdfbaa8693d7946b5e267