Pin the logstash and logstash-plugins everywhere they are installed
We are preparing to mirror the upstream opensearch debian repositories.

This will have the effect of making multiple versions of logstash and
logstash-plugins available for installation, with higher version numbers
automatically being selected.

Prior to this change, we wish to ensure that our logstash and
logstash-plugins packages are pinned to our current versions in puppet.

This will prevent any accidental package upgrades once we start syncing
the repos.

Hosts: C:logstash
Bug: T407199
Change-Id: I25768ef522ae4315d9a8690cd6387b7ea5b5de3c