sudoers: allow members of datacenter-ops group run homer
Currently only members of 'ops' have the rights to excute homer on
our cumin servers to configure network devices.

Only two members of dc-ops have this access, the rest are only part
of the datacenter-ops group.  The arrival of Nokia switchs, which
we configure using their JSON-RPC API, means we cannot use SSH to
configure switch interfaces, as the current configure-switch-
interfaces cookbook does.

This patch allows members of the datacenter-ops group to run homer
with sudo, which will allow them to run it when they connect
servers to the new switches.

Netops have discussed and are comfortable giving them the rights
to do this.

Bug: T402511
Change-Id: I6d42b2961e93b60f883dcc380bd0d1fdb8bf7297