url_downloader: remove hcaptcha proxy configuration
We have migrated these hosts to dedicated hcaptcha VMs, so this
configuration should be removed from the urldownloader hosts.

This commit will require a NOOP PyBal restart because the service
description has changed and our alerting is not smart enough to
distinguish between that and an actual code change.

Bug: T405631
Hosts: urldownloader1003.wikimedia.org
Change-Id: I08e685bc09121d1c405d3a7ddfcae316f798a433