gnmic: add collection for Nokia OSPF states
The new Nokia switches do not support the IETF OSPF-MIB so we can't
use our existing Icinga check to monitor if a link status has gone
down.

Unfortunately they also do not seem to support the openconfig OSPF
paths, so we cannot use the same paths as tested by Arzhel before
with Juniper.

This patch adds collection for some minimal status so we can see
the interface status (in terms of ospf), number of adjacencies
on it, and the status of any adjacencies.  This is really all we
need and will allow us to create alertmanager rules to signal if
there is a problem.

I ran a test instance this was the output from the prometheus
endpoint: https://phabricator.wikimedia.org/P84041

Hosts: netflow1002.eqiad.wmnet
Bug: T405558
Change-Id: Ia1bd3f82c513b21d4661f47a4c814c4a6fc99bfa