Configure reprepro to mirror upstream opensearch2 and opensearch3 repos
To date we have managed a local instance of an opensearch package
repository, into which we have added a single version of each of the
packages that we need.

Now we are moving to a different model whereby we will mirror the
upstream package repositories.

In preparation for this, we have pinned the logstash, opensearch, and
opensearch-dashboards packages.

Hosts: O:apt_repo
Bug: T407123
Change-Id: Ifbc89ff4c984ccbf6950f448edc241cbb3f56f0e