varnish: Add test for m.wikisource.org x-dt-host rewrite and POST
== m.wikisource.org ==

> regsub(req.http.host, "^([a-z0-9-]+\.)?m\.", …

It seems the only reason that the first match group is optional here
is to deal with m.wikisource.org. All other domains that start with
"m" are either mobile versions of MediaWiki wikis that use "www"
for their canonical domain, or static portal pages that likewise
use "www" as their canoical domain.

The only exception to my knowledge is m.wikisource.org. I'm tempted
to remove the question mark and optional nature, but the logic is
simpler and more robust if it keep it optional. For now let's just
add a test to make sure this doesn't regress.

== POST ==

In prep for redirecting test.m.wikipedia.org, add coverage for
POST requests as well.

Hosts: cp1102.eqiad.wmnet
Bug: T405931
Change-Id: I152309edf26b2502e0aeb1f2ee3543f6fc0ea3fb