varnish: add conditional to varnish::common::vcl for beta
The block in varnish::common::vcl that calls volatile for browser
detection does not apply to beta and therefore we add a conditional
around it to prevent Puppet breaking in the beta cluster.

Is this a good use of the etcd rules conditional? No, but for now, this
is what we are using to conditional on things that are not in beta. And
since there is some hint of unifying this work in Q4 of 2025-2026, the
idea is to keep all of this under one conditional so that later we can
replace it with something proper, like $supported_in_beta or similar.

Bug: T407966
Hosts: cp1100.eqiad.wmnet
Hosts: deployment-cache-text08.deployment-prep.eqiad1.wikimedia.cloud
Change-Id: Ic70ad532d0340796f77166477649837fa900fcd0