Update the apt components used for elasticsearch-curator
Although we have migrated to OpenSearch, we still require the use of
elasticsearch-curator to prune old indices, for the time being.

We forked the original elasticsearch-curator, so it is seems more
appropriate to give it a component/ prefix to its apt repository than a
thirdparty/ prefix.

We also need to move our currently hosted version to a new location in
the apt repository in preparation for the mirroring of upstream
opensearch repositories, in which it does not exist.

The thirdparty/elasticsearch-curator5 repository is no longer used by
any hosts, so we can remove it from the apt repository and remove its
upstream mirror component.

Note that this component will still require manual cleanup, as per:
https://wikitech.wikimedia.org/wiki/Reprepro#Removing_a_component

Hosts: O:apt_repo
Bug: T407199
Change-Id: Ia5bf9c590f6d732da5346886add6a21d303c0183