Add a thirparty/documentdb component to reprepro for trixie
We need to build a version of postgresql 17 that includes the documentdb
extension in order to work with FerretDB for GrowthBook.

The documentdb packages are licence compatible and are distributed by
FerretDB at: https://github.com/FerretDB/documentdb/releases/

We need to copy one of these deb packages to our local apt repositories,
which will subsequently be used in the container build pipeline for this
postgresql version.

Hosts: O:apt_repo
Bug: T408085
Change-Id: I260badde1b66dd5c1ec184f91db02a6f074811f3