site/role: create placeholder role/profile for tcpproxy
Creating a basic role and profile that installs just
the default haproxy package.

In bookworm and before we install from components and apparently
build our own packages.

But let's see how far we get with just the default that is in trixie
before we would have to build our own for this use case.

Bug: T408532
Change-Id: I0dbedfb994daa16272f65118f14b0254d62d2ccd