LVS: etcd data for druid-public-coordinator
We are adding the druid coordinator service to LVS as part of our ongoing work to use single urls
per service as opposed to using hardcoded hosts names which cause issues during maintenance
downtimes and general host issues.

This patch adds the etc-data for backend selection and for DNS discovery
required for the druid-public-coordinator.

Bug: T406222
Change-Id: I3a8861aa5a823bb165d434941ca9ba4d563b3fb9