gnmic: split bgp collection targets for Juniper/Nokia
Unfortunately pfw1-codfw has a problem with the gathering bgp stats
over gnmi if the "openconfig" prefix is requested.  However that is
needed for Nokia devices.

This is because of bugs on the versin of JunOS on that pair of
firewalls, which we already want to downgrade to reccomended version.

For now I need to split this into two groups so we can gather the BGP
stats from pfw1-codfw and alert if there is any issue.

Bug: T393996
Change-Id: I60f0e91e859b0e839d8906b37267aa8973c77708