P:toolforge::k8s::haproxy: Use hourly logrotate
Currently the HAProxy logrotate config is set to rotate daily, or
whenever the max size (2G) is reached, which usually happens once or
twice a day in addition to the rotation round at midnight. This patch
changes the rotation to happen at a more predicatable hourly interval
to support adding monitoring for this.

Bug: T408457
Change-Id: I799a3cb7f397ed9f7a52cb00fc3bc2236a6a6964