druid: Increase the size of the Druid broker cache size to 4GB
We have recently been hitting the druid broker cache size capacity much more frequetly.
This mainly happens a few days after being being cleared, presumably after a restart of each broker
process. Growing the cache size would reduce the amount of memory available for page-cache,
that druid uses heavily.
However, the total amount of memory available on the host is 128Gb so a 2GB increase is worth the
try and we shall monitor for any issues.

Bug: T408189
Change-Id: I11c2d1d919f3d3039b6f553d988dd743274fbfc4