global_config: add an urldownloader external service
This can be used to emit requests to external (read public) services
from a pod. We're planning to use this in an Airflow task that needs to
request https://feeds.spur.us/v2/anonymous-residential/latest to
retrieve data that we'll then inject into an opensearch server, as well
as to the data lake.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T408012
Hosts: O:deployment_server::kubernetes
Change-Id: I2fcb8115754857c1eb88e990cf5d105ba0758070
Hosts: O:deployment_server::kubernetes