toolforge::toolviews: Fix footgun with default values
These values are required in tools but not in toolsbeta. The previous
way of setting defaults avoids having to explicitely set dummy values in
toolsbeta, but it creates a situation where it's too easy to miss
situations where the values do need to be set (see T408454). Instead
replace the default values with Optional parameters in the profile, and
let the stricter types of the toolforge::toolviews class handle the real
validation when operating in the right project.

Change-Id: I6ddce7d8e140ad69232f76e65ee7ba326a6a6964