dnsrecursor: fix handling of auth_zones for wikimedia-common.yml.erb
Previously this was just a string, but in order to yaml it up it
needs to be a list of comma-delimited pairs.

The .erb template treated it as a list (but not a list of pairs)
so this patch has to change both the value and the handling.

Hosts: cloudservices1005.eqiad.wmnet, cloudservices2004-dev.codfw.wmnet
Bug: T381608
Change-Id: I801de1eafc0a4b99e2f8693f06ecee036f6c5bc4