dumps: Release the new MW Content File Export. Deprecate legacy XML dumps.
With this patch we release and start linking to the new MediaWiki Content File Exports.
We also deprecate the legacy XML Dumps, as these file exports are meant to replace them.

Additionally, we remove the public_mirrors.html file, as it was out of date.
Instead, we now redirect to the wiki page that host the same information but that is
kept up to date.

Hosts: O:dumps::distribution::server
Hosts: clouddumps1002.wikimedia.org
Bug: T401022
Bug: T408881
Bug: T414389
Change-Id: Ifa7c05b3b2ca40828b7d8f6725948b7e651365c8