toolforge: Handle interwiki redirects in front proxy
The iw tool was created in 2020 to support continued use of interwiki
links to Toolforge following the migration to per-tool *.toolforge.org
hostnames from the prior tools.wmflabs.org shared hostname. The
implementation of this tool which lasted for 5 years was a Kubernetes
Ingress object. Unfortunately newer versions of the nginx ingress system
used in Toolforge have lost support for the exact functionality needed.

In support of fewer places to look for config and a long ago promise in
Phabricator (T247432#6135378) that the iw redirector will be treated as
Toolforge infrastructure, the redirector functionality is finally being
added to the Toolforge front proxy.

The logic of the redirector is relatively straight forward. Incoming
requests look like `https://iw.toolforge.org/$tool/$path` where `$path`
may be empty. The response redirects to `https://$tool.toolforge.org/$path`.
Any query string attached to the inbound request should also be passed
as part of the redirect.

The inbound form allowed legacy redirects to remain unchanged in
wikitext by changing the target host in the interwiki map from
tools.wmflabs.org to iw.toolforge.org.

Bug: T247432
Bug: T408570
Change-Id: I9a4620ae3ce3b71dc6ce62132555ab4b05b35eee