deployment_server: default to PHP 8.3 in mwscript-k8s
This switches the default PHP version used (in the absence of explicit
selection via --php_version) to 8.3.

Like Ide16e37aea447285ec6ca261cd688d981d68087f, it logs a banner that
informs the user:
1. When they are being defaulted into 8.3 and how they can fall back to
   8.1 if they experience issues.
2. If they have explicitly selected 8.1, that the fallback option will
   be removed at a later date (so they should fix the problem).

I considered making this a bit more generic (e.g., so the banner logic
could be left in place, inactive when there's no migration), but I could
not come up with a simple-enough solution to warrant leaving it here for
a once-a-year event.

Bug: T405955
Change-Id: I82c88e17e7278364f5d8fb13c7c2b9e82767469d