Limit microcode installation to Bullseye
Starting with the Bookworm the CPU type is detected by the hw-detect
component of the Debian installer and amd64-microcode or intel-microcode
are installed accordingly. This also explains why we have amd64-microcode
installed on the handful of AMD machines we have despite it lacking support
in Puppet (which was the original reason why I started to investigate).

As such restrict the code to Bullseye hosts, once these are gone we can
remove it for good.

And also remove a stray comment related to mcelog, which is no longer use
in favour of rasdaemon.

PCC below for a Bullseye (cp), Bookworm (cumin) and Trixie (durum) host

Hosts: cp1103.eqiad.wmnet, cumin2002.codfw.wmnet, durum1001.eqiad.wmnet
Change-Id: I8bcbf3fe764677c194ce141eaeca494341f1ea1d