httpbb: adjust WDQS .git URL tests
Back in 2022 in If9f20664601a7de these tests were added with
an expected status code of 403, for being blocked actively.

At some point later the URLs disappeared entirely so now they are
404s.

Since a 404 is not an expected 403 that lead to failures in the tests.

Adjusting to reality to fix them.

Bug: T294917
Bug: T408744
Change-Id: I72a62254f7e0a2451617ab8cbfdee71218176cb3