gerrit: add discovery name as allowed destination range IPs for ssh
If we use the new gerrit.discovery.wmnet added in Iea0cf8d0cf5a4
in our HAproxy config as the backend name we also need to allow
it in the existing firewall rule which declares an allowed dest range.

Add the DNS name to Hiera (to be nice to things like test setups in
cloud) and look up both IPv4 and IPv6 in puppet.

Then add IPs to the existing firewall rule for port 29419 to gerrit SSH.

Bug: T365259
Change-Id: Ic56a8a52b44c452776090af7307350231f5b1e96