gerrit: add firewall rule to allow CDN caching servers to gerrit-ssh
A much simpler version of Ic56a8a52b44c452 that allows CDN caching
servers (and deployment hosts for testing) to connect to the gerrit
SSH port.

And only this, not allowing external users to connect to the port
on the gerrit host IP vs the service IP.

This is needed for proxy'ing gerrit-ssh via the new tcp-proxy VMs.

Bug: T365259
Change-Id: If6209cd9d951e7a05009494f4d2b4b1ae4aee0bd