gerrit: allow production networks to connect to gerrit-ssh
follow-up to If6209cd9d951e7a05

We need to allow the TCP-proxies to connect go gerrit-ssh
and the CDN caches to the tcp-proxies.

Not CDN caches directly to gerrit.

Upon further thought/discussion it is fine to allow PRODUCTION_NETWORKS.

This service is available to the world regardless
and it means we don't have to maintain a list of TCP proxy servers
or add them as a global networking constant.

Bug: T365259
Bug: T408532
Change-Id: If64c00edb994bd22578b6327cc4966e363e32bdd