P:cache::haproxy: ensure x-requestctl is updated
Previously, the x-requestctl request header consumed by upstream varnish
was only updated in requestctl-filter backends.

That works perfectly well *prior* to introducing known-client DSL, but
now leaves an awkward gap. If identification is enabled, we won't select
any of the filter backends, and lose any known-client-related logging to
txn.x_requestctl when it round-trips through varnish via x-analytics.

See [0] for a longer-form description of the problem.

In any case, this now updates x-requestctl when txn.x_requestctl has
been set in the no-filter-backend case as well.

This should be a noop for all previously existing use cases, since
known-client DSL is the only one that now touches txn.x_requestctl in
the listen section.

[0] https://phabricator.wikimedia.org/T403220#11342816

Bug: T403220
Change-Id: I6fa4c196afbf4c68af35d201e200b975f68de37a