dumps: Fix missing trailing slash in cirrus-search-index path
Without the trailing slash it creates the directory in the
target, but we only want to sync the contents, not the named
directory itself.

Will also require an SRE to cleanup the directory that was
created by the initial deployment without this fix.

Bug: T366248
Hosts: clouddumps1001.wikimedia.org,clouddumps1002.wikimedia.org
Change-Id: I4df80bf43f037f293fdd8774887ab87f84087555