deployment_server: migrate mw-wikifunctions to PHP 8.3
This service uses single-version images / routing, which in practice
means that all traffic is served by the group1 release.

Together with the fact that it does not receive enough traffic for an
incremental capacity-based migration to prove useful, and that at this
point we consider the switch to 8.3 to be low-risk, we'll switch in one
step and monitor metrics / logs.

A similar procedure was used during the 8.1 migration.

Bug: T405955
Change-Id: Ia711395019c080edb102e4514161a21d1aca3a6c