w[cd]qs: Log more provenance headers
In addition to superset and turnilo we would like to add the `x-ja3n` `x-is-browser` and
`x-is-client-ip` to the nginx logs.

This should help us identify bot traffic in the future.

See also I4ebcd48a128f07a71d3d949992ae2642987ead22 , which
created the file resource but did not actually apply it.

Hosts: wcqs2001.codfw.wmnet,wdqs2007.codfw.wmnet
Bug: T408123
Change-Id: Iba64494fd1c284ced69b7bfabb69c50ebbf3ce31