P:wmcs::cloud_private_subnet: Support enabling jumbo frames
We want to support jumbo frames on most WMCS hardware, and on both the
base (cloud-hosts*) interface for Ceph reasons and for the cloud-private
interface for Neutron reasons. So even if bumping it on the base
interface is not strictly related to the cloud-private setup, I felt
that for consistency it is better to have both set in the same place
rather than add the cloud-hosts override all over the place.

And enable it for testing on cloudvirt2004-dev.

Bug: T330075
Change-Id: I136bef6f2f5d07172a99c19b3c894a146a6a6964