envoyproxy: Envoy config updates for 1.32
The prior config is valid up through 1.32, this just removes warnings
from the log and unblocks the next upgrade.

Use a downstream connections resource monitor[1] to limit total
connections, instead of the deprecated[2] runtime key we're currently
using.

As a result, drop the hardcoded static runtime layer in the bootstrap
config (since this was the only thing it was used for) and replace it
with a hardcoded overload manager.

[1]:
https://www.envoyproxy.io/docs/envoy/v1.32.12/api-v3/extensions/resource_monitors/downstream_connections/v3/downstream_connections.proto.html
[2]: https://www.envoyproxy.io/docs/envoy/latest/version_history/v1.30/v1.30.0#deprecated

Bug: T409510
Change-Id: Iaf8deac882ab83f15bb0e74e6a4230316a6a6964