Standardize the group ownership of the keytab files on stat servers
We have observed that the group ownership of three keytabs is different
from the standard configuration on three of the kerberos keytabs that
are deployed to the stat servers.

This commit brings those three files into line with the other keytabs,
by resetting their group ownership back to match the primary owner
entry.

Hosts: O:statistics::explorer
Bug: T409770
Change-Id: Id5257f87cbfd037f20947109aa935411af46ba6d