toolsdb: increase innodb_log_file_size to 512M
This should make it less likely for mariadb to fail during recovery
after a crash. The main downside is slower recovery time, but recoveries
have always been quite fast so far, it's ok if they get a bit slower.

We might be able to increase this value even more, but I would like to
try a smaller increase first.

Bug: T409922
Change-Id: I5815f79f6aa16c6081513a497504833d1b145f67