homer/puppetmaster: Make sure the commitmsg hook does not double-add user
When adding the user prefix to the commit message, the commitmsg hook
would double it up if it was already there (if the user added it
themselves). E.g. 123ffeedbc91402fa69f0e7409e882b570f993f2 on the
private repo. Change the RE of the sed call to make sure it's there only
once (in essence, overwrite it with a duplicate if it's already there).

Examples:

before:
$ echo foo message |sed -e "1s/^/(${SUDO_USER}) /"
(klausman) foo message
$ echo "(klausman) foo message" |sed -e "1s/^/(${SUDO_USER}) /"
(klausman) (klausman) foo message

after:
$ echo foo message |sed -re "1s/^(\($SUDO_USER\) |)/(${SUDO_USER}) /"
(klausman) foo message
$ echo '(klausman) foo message' |sed -re "1s/^(\($SUDO_USER\) |)/(${SUDO_USER}) /"
(klausman) foo message

Any prefix that isn't exactly what the sed call would add would be
left-as is and the normal prefix added.

Change-Id: Icd375eeb39d79fddd9f6b638a6dfdd0c3637dfe9