admin: upgrade lpintscher from ldap_only to analytics-privatedata
Lydia Pintscher has had shell access in the past, then did not
need that anymore and became an "ldap_only" user.

Now requests to be added to analytics-privatedata-users.

So re-activating the "shell" user but without an SSH key
for the standard access to private data.

Removing the "ldap_only" section.

Bug: T409933
Change-Id: I96a88152ed65c6977bf70d7f2f7d75a509eba2e8