wmcs k8s nfs: add NFS tracing script
Why:
As part of T399313 we wish to track tools interpendencies on toolforge.
This patch aim at tracing the NFS traffic made by tools storing the logs
in the new infra-tracing Loki instance.

What:
Add the Python script that will perform the tracing and will push the
logs to Loki every X log lines or Y seconds (configurable).
Add also the related bits to manage it as a systemd unit.

Hosts: auto
Bug: T399313
Change-Id: Iaa6993a5803941700659cf19944c3e189471ffff