partman: Apply workarounds to swap handling affecting trixie installations
With the current DB partman recipe, the creation of the swap partitions
fails with a pre-existing swap partition:

The swap partition is correctly parsed and detected by d-i, but
ultimately when the d-i component responsible for setting up swap
(partman-basicfilesystems) attempts to create the swap partition
using "mkswap /dev/sda2" this fails with

mkswap: write error: Text file busy

And in fact the existing swap partition is activated:

~ # cat /proc/swaps
Filename  Type      Size    Used Priority
/dev/sda2 partition 7812092 0   -2

Possibly something changed in an earlier part of d-i or maybe the
initrd, so that a pre-existing swap partition from an earlier
installation is now activated.

This is also specific to the DB.cfg partition, we haven't seen
this with a different partman config on trixie. I'm going to
report this to Debian, but as a workaround and to unblock DB
installations on trixie, this adds two workarounds, which
have also been verified to work with Bookworm:

- db.cfg has been extended to skip the confirmation dialogue in
  d-i (which is perfectly fine, since we can simply reuse
  the existing swap partition, the format hasn't changed since
  Linux 1.x)
- reuse-db.cfg has been extended to suppress a warning about a
  missing swap partition (which is also correct since the reuse
  config doesn't configure swap)

Bug: T408777
Change-Id: Ie5ac52e6bcb5dd48fb39d99f5d1951108053ef58