releases: control jenkins service by host name
follow-up to Ia3d4dd513dd0916356

(5fb233ae744, I983dca5b73a1d23)

Still confirming if the Hiera lookups work as intended or not.

It appears that using role/eqiad or role/codfw is not working
as expected.

Setting on host name level as it was before today and also making
the switch to codfw.

Bug: T392127
Bug: T391578
Change-Id: I65c02da3ccf379ad1f5fcf4e102a962a9d656782