wdqs: Do not create task on failure of the WDQS LDF endpoints
The WDQS LDF endpoints don't have a formal SLO, are backed by a single
server, and are expected to fail reasonably often. Creating a task
each time this endpoint is down is counter productive. Changing the
severity of that check to "warning" should make those alerts less noisy.

More info on:
* https://wikitech.wikimedia.org/wiki/Wikidata_Query_Service#LDF_(Linked_Data_Fragments)_endpoint
* https://www.wikidata.org/wiki/Wikidata:Data_access#Linked_Data_Fragments_endpoint

Bug: T408853
Change-Id: I33512d3e61720f7e7d1928cba9c25144efe662a3