O:haptcha::proxy: add new role for hCaptcha proxy VMs (bird/anycast)
This is mostly a duplication of O(hcaptcha_proxy), but adapted for the
new bird/anycast setup described in T409780. The reason for keeping this
separate is that we still want to preserve the existing setup in place,
and only switch over hCaptcha once the rollout of the new bird/anycast
proxy has been tested.

Until that happens, we will keep both systems running, after which we
will deprecate the old setup.

(We also set routed Ganeti for magru/esams.)

Bug: T409780
Change-Id: Ie05a80afb7c8329afe53016614a3e9886a67f4d4