lvs1020: move row C vlans to primary and add new C/D per-rack vlans
This patch adjusts the network configuration for lvs1020 to connect
to the existing row C vlans over its primary link (which is in row
D, but once moved to Nokia switch will also be able to reach row
C on them).

Additionally it adds the new per-rack vlans for rows C and D to the
same primary uplink so they are in place for when those vlans go
live.

Bug: T405609
Change-Id: I1d5223f0f3749b9c821d2082cef88006ec23576f