lvs1019: move row D vlans to primary and add new C/D per-rack vlans
This patch adjusts the network configuration for lvs1019 to connect
to the existing row D vlans over its primary link (which is in row
C, but once moved to Nokia switch will also be able to reach row
D on them).

Additionally it adds the new per-rack vlans for rows C and D to the
same primary uplink so they are in place for when those vlans go
live.

Bug: T405628
Change-Id: Ie4d57fcfbe8585e2d323251c0caedb330080de2a