relforge: Change to test role
The relforge instances are being temporarily repurposed to test
opensearch 3 functionality in preperation for a future upgrde.
These instances are already test instances, they are not in the
pathway of any production code. This test role is based on
role::test, but modified to ensure search platform can adjust the
instance as needed.

This is being migrated to a test role, rather than
providing a puppetization of opensearch 3, because the production
instances all need to migrate 1->2->3 in sequence, but here
we want to test the functionality that we get by moving to 3 to
provide justification for the future update.

Bug: T410681
Hosts: relforge1008.eqiad.wmnet
Change-Id: Ib44c72e146210e5466cb604c57d0267c7769e32b