admin: remove wmde-fisch from releasers-wikidiff2
As part of an audit of releasers-* shell groups:

Christoph has said he and his team have not been working on this
library for some time and are not maintaining it.

So removing him from the access group.

It remains to be seen if this means we remove the entire group.

And whether we should keep historic releases around forever (but why not,
right?).

Change-Id: I8bb57740471f18bbff0ec6c77bb07fa4c4cd578c