deployment_server: switch mw-script/main to PHP 8.3
This switches the main release over to 8.3, now that support for 8.1
fallback in mwscript-k8s has been dropped.

The next patch in this series will switch mwscript-k8s over to following
main once again.

In contrast to when we cleaned up after the 8.1 migration, we'll retain
the next release in config, which is more consistent with how we manage
next / migration for other deployments and (IMO) simplifies this switch
process.

Bug: T405955
Change-Id: I45e3e703ac34ebb4355f809a97d22867c204c36a