releases::mediawiki: change the time when jenkins is restarted
As reported on T410729 the jenkins service on releases-jenkins
(releases*) gets auto restarted by wmf_auto_restart daily.

The time when this restart happens is randomized unless a specific
hour and minute is given:

----
# This define can be used to add an automatic restart for a stateless service.
# wmf-auto-restart checks whether any dependent library has been refreshed and
# if that's the case, a restart is triggered. The restarts are spread out over
# the course of the day via fqdn_rand() if not set explicitly using $restart_hour
# and $restart_minute.
----

As Jaime explains the value it picked on the new host releases2003
is at 00:29 and too close to the nighly builds running at 00:30.

Set the time to 8 minutes after 8 instead.

Bug: T410729
Change-Id: Id8187097dff7148073b0ad188f2e9c4cc1b91261