alertmanager: Add a receiver for experiment-platform-ircmailtask
We'll be setting up SLO-based alerts on a, well, experimental basis.
This config might change down the line.

As an initial setup, by the team's request (T398869#11368002) we'll
route SLO alerts to *all* of:

- #wikimedia-operations on IRC
- #wikimedia-data-platform-alerts on IRC
- data-engineering-alerts@wm.o by email
- tasks in the Experiment Platform project on Phab

Pyrra emits SLO alerts with severity "warning" or "critical" (depending
on the window and burn rate) so route both to the -ircmailtask receiver
for now -- down the line, that might change too. For now, that also
doesn't conflict with any other alerting sources, as nothing else is
using this team label except with severity=task.

Bug: T398869
Change-Id: Ie94d05de167cb354801bc96e7630f77b6a6a6964