profile::thanos::swift: add tegola account for staging
There is a new maps beta/staging stack currently being built
to test new and invasive changes in the future. In order to
avoid impacting any cached production tile, it makes sense
to create a new separate account for tegola staging.

Add a new account for tegola staging to the Thanos Swift's config
(the private data will be added before this change).

Bug: T409528
Change-Id: I365be7ed803225db255105cc49e38cd3b29ec808