admin/releases: deprecate the releasers-wikibase shell user group
We have done an audit of releasers-* shell groups and as part of
this we received feedback that WMDE is not going to release wikibase
tarball releases anymore in the future.

These releases have been replaced by docker distro.

Therefore the shell user group can be deprecated and removed
from the releases hosts.

File ownership moves to root:root. Files will stay up in archive mode.

Bug: T410418
Change-Id: Id13756caa2d3de5964c8b88839434012d9931995