varnish: change contiditon for generating scoped file rules
We don't want the confd file to only be present when the rate limiting
for known bots is enabled, because that might lead to a failing varnish
reload if the file is missing but referenced in VCL, and the ordering
of the confd::file resource and the varnish changes is not guaranteed.

Same goes for the authorization scoped rules.

Change-Id: Ib2f7ff5f64a5da5171256dc16ba0eb94f8deda0e