backup: Increase the number of maximum storage for repos to 50 TB
Repo backups were limited to 25TB or 50 Volumnes. Increase that
limit to 100 volumes or 50TB in total, as we had reached the usage
limit, and backups were blocked.

After deploy, this usually requires a config reload on bacula
director to transmit it to the storages, and then a bacula-storage
daemon restart or reload.

Change-Id: I4d8d3b83773eef05405f5e8c11c2b16d07ad5e00