toolforge: add ingress for infra-tracing-loki
What:
Make the toolforge k8s haproxy listening on port 30004 to allow access
with an svc URL to the infra-tracing-loki instance in TCP mode.
The TLS termination is done by the Loki Nginx gateway.
To be used by WMCS's grafana to access Loki's datasource.

The related CNAME patch is:
https://gitlab.wikimedia.org/repos/cloud/toolforge/tofu-provisioning/-/merge_requests/103

Hosts: auto
Bug: T399313
Change-Id: I1ac308793dfaa6180246d27789c9b2af3456983e