firewall: Declare resources for both providers
On Cloud VPS, any use of firewall::service seems to be applied before
profile::firewall (and thus class[firewall]) could be applied. As
firewall::service relies on variables set in the firewall class, this
results in a duplicate declaration error.

This patch works around that by having firewall::service apply the
resources for both firewall types (nftables and ferm). Both drivers
use virtual resources and will do nothing if the base class is not
present. (For example this is how, so far, declaring a ferm::service on
a Cloud VPS VM that doesn't run Ferm had no effect.)

Bug: T411089
Hosts: auto
Change-Id: If37a76dc3b676b2c9f28ca8e2617745c6a6a6964