trafficserver: Add missing REST Gateway for Beta Cluster
We need to write and maintain something like for Beta Cluster
either way. I suggest we it in a way that maps directly to
to a backend service, rather than via either RESTBase or
REST Gateway.

This means we don't add transparent Envoy features like Prometheus
metrics and connection pooling, but that seems worth the trade-off
in reduced updates, breakages, and other maintenance.

While at it:

* Update docs of prod version to explain where rest-gateway is.

* Switch return 0 to TS_LUA_REMAP_NO_REMAP (=== 0) for clarity.

* Use `orig_path` var name for consistency with the prod version.

* Switch remnant beta.wmflabs.org host for beta.wmcloud.org equivalent.
  This was found as part of the new unit test.

Bug: T404387
Change-Id: If1fc463d2e9ba90744375fe56e529dcf59077662